requirements.txt
# Dependencias mínimas para DEPLOY (fastapi_app_onnx.py)
# Para desarrollo local, usar: pip install -r requirements_local.txt
# Tamaño total: ~150 MB (optimizado para Vercel/Railway/Render)

# ═══════════════════════════════════════════════════════════════════════
# CORE - FastAPI & Web Server
# ═══════════════════════════════════════════════════════════════════════
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0

# ═══════════════════════════════════════════════════════════════════════
# DATA & ML - Mínimo necesario
# ═══════════════════════════════════════════════════════════════════════
numpy==1.24.3

# ═══════════════════════════════════════════════════════════════════════
# ONNX - Solo Runtime (NO conversion tools)
# ═══════════════════════════════════════════════════════════════════════
onnxruntime==1.16.3

# ═══════════════════════════════════════════════════════════════════════
# HTTP CLIENT - Para cargar modelos desde Vercel Blob
# ═══════════════════════════════════════════════════════════════════════
requests==2.31.0

# ═══════════════════════════════════════════════════════════════════════
# PRODUCCIÓN (opcional - descomentarlo para usar Gunicorn)
# ═══════════════════════════════════════════════════════════════════════
# gunicorn==21.2.0
